+++ CHANGELOG.md
@@ -8,6 +8,7 @@ AdminNeo 5.1.0
 
 - Add defaultServer and defaultDatabase configuration parameters
 - Allow creating generated columns (by @vrana)
+- MySQL: Display converting function for binary, bit or geometry fields (by @vrana)
 - MS SQL: Add support for computed columns (by @vrana)
 - SQLite: Add support for generated columns (by @vrana)
 - PostgreSQL: Add support for indexes on materialized views (by @vrana)